requirements.txt
Flask
Pillow
requests
flask[async]
psycopg2
minio
Werkzeug